Moving disk based aggregation processor before merging in the mem based one
